requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
idna==3.3
libretranslatepy==2.1.1
lxml==4.8.0
lyricsgenius==3.0.1
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.1
translate==3.6.1
urllib3==1.26.8
googletrans~=3.1.0a0